# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Portal de Autoatendimento de TI",
    "description": "API do MVP academico. Ambiente de demonstracao; dados e integracoes sao ficticios.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/tickets": {
      "post": {
        "tags": [
          "chamados"
        ],
        "summary": "Open Ticket",
        "operationId": "open_ticket_api_v1_tickets_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chamados"
        ],
        "summary": "Get Tickets",
        "operationId": "get_tickets_api_v1_tickets_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketPriority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketCategory"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "assignee_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Assignee Id"
            }
          },
          {
            "name": "unassigned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Unassigned"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "deprecated": true,
              "title": "Status Filter"
            },
            "deprecated": true
          },
          {
            "name": "priority_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketPriority"
                },
                {
                  "type": "null"
                }
              ],
              "deprecated": true,
              "title": "Priority Filter"
            },
            "deprecated": true
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TicketSummary"
                  },
                  "title": "Response Get Tickets Api V1 Tickets Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tickets/{ticket_id}": {
      "get": {
        "tags": [
          "chamados"
        ],
        "summary": "Get Ticket",
        "operationId": "get_ticket_api_v1_tickets__ticket_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tickets/{ticket_id}/assignment": {
      "patch": {
        "tags": [
          "chamados"
        ],
        "summary": "Patch Assignment",
        "operationId": "patch_assignment_api_v1_tickets__ticket_id__assignment_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketAssignmentUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tickets/{ticket_id}/priority": {
      "patch": {
        "tags": [
          "chamados"
        ],
        "summary": "Patch Priority",
        "operationId": "patch_priority_api_v1_tickets__ticket_id__priority_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketPriorityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tickets/{ticket_id}/status": {
      "patch": {
        "tags": [
          "chamados"
        ],
        "summary": "Patch Status",
        "operationId": "patch_status_api_v1_tickets__ticket_id__status_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "ticket_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Ticket Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TicketAssignmentUpdate": {
        "properties": {
          "assignee_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignee Id"
          }
        },
        "type": "object",
        "required": [
          "assignee_id"
        ],
        "title": "TicketAssignmentUpdate"
      },
      "TicketCategory": {
        "type": "string",
        "enum": [
          "ACCESS",
          "SOFTWARE",
          "NETWORK",
          "HARDWARE",
          "SECURITY",
          "OTHER"
        ],
        "title": "TicketCategory"
      },
      "TicketCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 5,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 10,
            "title": "Description"
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "category"
        ],
        "title": "TicketCreate"
      },
      "TicketDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "requester": {
            "$ref": "#/components/schemas/UserRef"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TicketEventRead"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "id",
          "protocol",
          "title",
          "category",
          "priority",
          "status",
          "requester",
          "assignee",
          "created_at",
          "updated_at",
          "resolved_at",
          "description"
        ],
        "title": "TicketDetail"
      },
      "TicketEventRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "author_id": {
            "type": "string",
            "format": "uuid",
            "title": "Author Id"
          },
          "author": {
            "$ref": "#/components/schemas/UserRef"
          },
          "from_status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TicketStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "to_status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "author_id",
          "author",
          "from_status",
          "to_status",
          "comment",
          "created_at"
        ],
        "title": "TicketEventRead"
      },
      "TicketPriority": {
        "type": "string",
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "TicketPriority"
      },
      "TicketPriorityUpdate": {
        "properties": {
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          }
        },
        "type": "object",
        "required": [
          "priority"
        ],
        "title": "TicketPriorityUpdate"
      },
      "TicketStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "TRIAGE",
          "IN_PROGRESS",
          "RESOLVED"
        ],
        "title": "TicketStatus"
      },
      "TicketStatusUpdate": {
        "properties": {
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "TicketStatusUpdate"
      },
      "TicketSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "protocol": {
            "type": "string",
            "title": "Protocol"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "category": {
            "$ref": "#/components/schemas/TicketCategory"
          },
          "priority": {
            "$ref": "#/components/schemas/TicketPriority"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "requester": {
            "$ref": "#/components/schemas/UserRef"
          },
          "assignee": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "protocol",
          "title",
          "category",
          "priority",
          "status",
          "requester",
          "assignee",
          "created_at",
          "updated_at",
          "resolved_at"
        ],
        "title": "TicketSummary"
      },
...
      "UserRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "UserRef"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Informe somente o token retornado por POST /api/v1/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}